refactor: support iife
