+++ src/ast/variables/Variable.ts
@@ -4,11 +4,9 @@ import type { RenderOptions } from '../../utils/renderHelpers';
 import type { HasEffectsContext } from '../ExecutionContext';
 import type { NodeInteraction } from '../NodeInteractions';
 import { INTERACTION_ACCESSED } from '../NodeInteractions';
-import type CallExpression from '../nodes/CallExpression';
 import type Identifier from '../nodes/Identifier';
-import type SpreadElement from '../nodes/SpreadElement';
 import { ExpressionEntity } from '../nodes/shared/Expression';
-import type { ExpressionNode, NodeBase } from '../nodes/shared/Node';
+import type { NodeBase } from '../nodes/shared/Node';
 import type { VariableKind } from '../nodes/shared/VariableKinds';
 import type { ObjectPath } from '../utils/PathTracker';
 
@@ -38,13 +36,9 @@ export default class Variable extends ExpressionEntity {
 	addReference(_identifier: Identifier): void {}
 
 	onlyFunctionCallUsed = true;
-	argumentsList: (ExpressionNode | SpreadElement)[][] = [];
 	addUsedPlace(usedPlace: NodeBase): void {
 		const isFunctionCall = usedPlace.parent.type === 'CallExpression';
-		if (isFunctionCall) {
-			const callExpression = usedPlace.parent as CallExpression;
-			this.argumentsList.push(callExpression.arguments);
-		} else {
+		if (!isFunctionCall) {
 			this.onlyFunctionCallUsed = false;
 		}
 	}
